% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_aggregate_counts.R
\name{alt_aggregate_counts}
\alias{alt_aggregate_counts}
\title{Alternate Aggregate UCLA data for website groupings}
\usage{
alt_aggregate_counts(
  date_cutoff = DATE_CUTOFF,
  all_dates = FALSE,
  week_grouping = TRUE
)
}
\arguments{
\item{date_cutoff}{date, the earliest date of acceptable data to pull from, 
ignored if all dates is true}

\item{all_dates}{logical, get time series data rather than just latest counts}

\item{week_grouping}{logical, use weekly grouping for past data? else monthly}
}
\value{
data frame with aggregated counts at the state level by web groups
}
\description{
Reads the UCLA aggregates counts for states for most recent
data after a given date cutoff and reports either state level data or national
data. States include values for the 50 states broken down by carceral type,
prison, ICE, Federal, Juvenile, Psychiatric, and county. For prisons, data
from the Marshall project is also incorporated.
}
\examples{
\dontrun{
alt_aggregate_counts()
}
}
