% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_web_group.R
\name{assign_web_group}
\alias{assign_web_group}
\title{Assigns UCLA website groupings}
\usage{
assign_web_group(dat)
}
\arguments{
\item{dat}{data frame with columns for Jurisdiction and Age (at least)}
}
\value{
original data frame with new Web.Group column added
}
\description{
Assigns UCLA website groupings based on a combination of a Jurisdiction and 
Age field. Web groupings include ICE, Federal, Juvenile, Prison, Psychiatric, 
and County. ICE and Federal facilities will never be assigned Juvenile (i.e., 
Jurisdiction takes precedence for ICE and Federal facilities, and Age takes
precedence for State, County, and Psychiatric facilities).
}
\examples{
assign_web_group(
    tibble(
        Name = "Sample youth facility", 
        Age = "Juvenile", 
        Jurisdiction = "state"
    ) \%>\% add_row(
        Name = "Sample federal facility", 
        Age = "Adult", 
        Jurisdiction = "federal"
        )
)
}
