% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fac_col_txt.R
\name{clean_fac_col_txt}
\alias{clean_fac_col_txt}
\title{Cleans character data that has been scraped from websites}
\usage{
clean_fac_col_txt(x, to_upper = FALSE)
}
\arguments{
\item{x}{character vector to clean}

\item{to_upper}{logical, convert characters to upper case}
}
\value{
cleaned character vector
}
\description{
Cleans data usually associated with facilities or column names. Specifically
removes special characters (except spaces and dashes), capitalizes COVID, standardizes
COVID-19 references, removes new line indicators and excess white space,
and changes "state-wide" to "statewide" (case-insensitive).
Optionally capitalizes output.
}
\examples{
clean_fac_col_txt(" Messy string \n\r data   ")

}
