% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_vector_warn.R
\name{coalesce_vector_warn}
\alias{coalesce_vector_warn}
\title{A re-coding of the coalesce_vector function to include warnings when multiple
values are given which are not NA and are different}
\usage{
coalesce_vector_warn(x)
}
\arguments{
\item{x}{vector to coalesce}
}
\value{
vector of length 1 of coalesced values
}
\description{
A re-coding of the coalesce_vector function to include warnings when multiple
values are given which are not NA and are different
}
\examples{
coalesce_vector_warn(c(1,2))

}
