% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_with_warnings.R
\name{coalesce_with_warnings}
\alias{coalesce_with_warnings}
\title{A re-coding of the coalesce function to include warnings when multiple
values are given which are not NA and are different}
\usage{
coalesce_with_warnings(...)
}
\arguments{
\item{...}{vectors of equal length and type to coalesce}
}
\value{
vector of coalesced values
}
\description{
A re-coding of the coalesce function to include warnings when multiple
values are given which are not NA and are different
}
\examples{
coalesce_with_warnings(c(1, 2, 3), c(4, 5, 6))
coalesce_with_warnings(c(1, 2, 3), c(1, 2, NA))

}
