% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_roll_sum.R
\name{diff_roll_sum}
\alias{diff_roll_sum}
\title{A function to calculate a rolling sum of first differences}
\usage{
diff_roll_sum(x, date_vec = NULL, window = 14)
}
\arguments{
\item{x}{a numeric vector}

\item{date_vec}{date, vector of dates of equal length to x default NULL}

\item{window}{int, integer indicating rolling sum window length}
}
\description{
Calculates a rolling sum of first differences on a vector, or optionally and
recommended, calculates first differences and rolling sum by day where
missing day values are treated as the same as the last observed value.
}
\examples{
\dontrun{
read_scrape_data(all_dates = TRUE, state = "North Carolina") \%>\%
    filter(Name == "NORTH CAROLINA CORRECTIONAL INSTITUTION FOR WOMEN") \%>\%
    mutate(Res.Act.Est = diff_roll_sum(Residents.Confirmed, Date)) \%>\%
    ggplot(aes(
        x = Date, y = Res.Act.Est, color = Name, fill = Name)) +
    geom_line(size = 1.5) +
    geom_area(alpha = .5) +
    theme_behindbars() +
    scale_color_bbdiscrete() +
    scale_fill_bbdiscrete() +
    theme(legend.position = "none") +
    labs(y = "Estimated Active Cases") +
    ggtitle("Monitoring Facility Outbreaks")
}

# example of difference using position vs using dates for rolling sum diffs
x <- c(1, 1, 4, 4, 7, 9)
sd <- Sys.Date()
date_x <- sd - c(10, 9, 8, 6, 4, 2)

diff_roll_sum(x, window = 2)
\dontrun{
# returns a warning
diff_roll_sum(x, date_x, window = 2)
}

}
