% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTable.R
\name{ExtractTable}
\alias{ExtractTable}
\title{Function to extract all tables from an image}
\usage{
ExtractTable(img, api_key = NULL)
}
\arguments{
\item{img}{character indicating file location of image or an object of
class 'magick-image'}

\item{api_key}{character extractable api key. If NULL will search Environment
variables for object 'EXTRACTABLE_API_KEY'}
}
\value{
list of matrices and vectors of extracted table information
}
\description{
Extracts tables from an image, returns a list of matrices and vectors with
extracted information.
}
\examples{

"https://dps.hawaii.gov/wp-content/uploads/2020/03/" \%>\%
    paste0("Inmate-Test-Report-12.1.20.jpg") \%>\%
    magick::image_read() \%>\%
    ExtractTable()
}
