% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genstate_vax.R
\name{get_genstate_vax}
\alias{get_genstate_vax}
\title{Get general population state COVID vaccine data from CDC and Census 2020 18+ pop estimates}
\usage{
get_genstate_vax()
}
\value{
a data frame with the following columns: State, Date,
General.Initiated, General.Adult.Population
}
\description{
Pulls the latest data from the CDC on COVID vaccinations initiated
for adults in a state and then merges alongside a population number from the 18+ 2020
Census population estimates for that state to use for rate calculation.
}
\examples{
\dontrun{

# get state-level vax data
st_gen_df <- get_genstate_vax()


}
}
