% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric_description.R
\name{get_metric_description}
\alias{get_metric_description}
\title{Gets the description for a given COVID-19 metric}
\usage{
get_metric_description(metric, short = FALSE)
}
\arguments{
\item{metric}{character string of the given metric}

\item{short}{logical, whether to return the short (vs. long) description}
}
\value{
character string of the given metric's description
}
\description{
Given a COVID-19 metric collected in the scraped data, returns a description of 
that metric, primarily useful for plotting (e.g. axis labels, titles, etc.).
}
\examples{
get_metric_description("Residents.Active")

}
