% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_src_by_attr.R
\name{get_src_by_attr}
\alias{get_src_by_attr}
\title{Gets a web element based on a specific attribute that matches regex}
\usage{
get_src_by_attr(
  base,
  css,
  xpath,
  attr,
  attr_regex,
  date_regex = NULL,
  date_format = "mdy"
)
}
\arguments{
\item{base}{character string of the URL}

\item{css}{css option to search html nodes}

\item{xpath}{xpath option to search html nodes}

\item{attr}{character the attribute to pull a string from}

\item{attr_regex}{character regex to search in attributes}

\item{date_regex}{character string to extract date from attribute}

\item{date_format}{character the format to convert the date from}
}
\value{
character of the url specified by the search
}
\description{
Gets a web element based on a specific attribute that matches regex
}
\examples{
\dontrun{
get_src_by_attr(
    base = "https://doc.arkansas.gov/covid-19-updates/",
    css = "a",
    attr = "href",
    attr_regex = "(?i)stats.?update",
    date_regex = "\\\\d+-\\\\d+-\\\\d+")
}

}
