% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_coalesce.R
\name{group_by_coalesce}
\alias{group_by_coalesce}
\title{A all in one group by and coalesce function}
\usage{
group_by_coalesce(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{...}{one or more unquoted expressions separated by commas. Variable
names can be used as if they were positions in the data frame,
so expressions like x:y can be used to select a range of variable}
}
\value{
vector of coalesced values
}
\description{
Groups by and coalesces the remaining columns such that only one value
is returned per column. If multiple values are present that are not NA
a warning is returned with the offending group, column name, and values.
}
\examples{
df_safe <- data.frame(
    A=c(1,1,2,2,2),
    B=c(NA,2,NA,4,4),
    C=c(3,NA,NA,5,NA),
    D=c(NA,2,3,NA,NA),
    E=c(NA,NA,NA,4,4))

df_warn <- data.frame(
    A=c(1,1,2,2,2),
    B=c(NA,2,NA,4,4),
    C=c(3,NA,NA,5,NA),
    D=c(NA,2,3,NA,NA),
    E=c(NA,NA,NA,4,5))

group_by_coalesce(df_safe, A)
group_by_coalesce(df_warn, A)

}
