% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fac_name.R
\name{is_fac_name}
\alias{is_fac_name}
\title{Returns TRUE if a facility name exists in the facility crosswalks}
\usage{
is_fac_name(
  fac_name,
  state,
  fac_info = NULL,
  fac_spellings = NULL,
  include_alt = T
)
}
\arguments{
\item{fac_name}{character string of the facility to plot}

\item{state}{character string of the facility's state}

\item{fac_info}{data frame with facility_info sheet}

\item{fac_spellings}{data frame with facility_spellings sheet}

\item{include_alt}{logical, whether to include alternative spellings}
}
\value{
logical, TRUE if fac_name is a valid name and FALSE otherwise
}
\description{
Returns TRUE if a facility name exists in the facility crosswalks, FALSE 
otherwise. Function is case-insensitive, and strips special characters, new line 
indicators, and excess white space (per behindbarstools::clean_fac_col_txt).
}
\examples{
is_fac_name("Los Angeles Jails", "California")

}
