% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_federal.R
\name{is_federal}
\alias{is_federal}
\title{Simple function that returns TRUE/FALSE if "federal" is detected, case insensitive}
\usage{
is_federal(x, ...)
}
\arguments{
\item{x}{string}

\item{...}{inherited argument from str_detect}
}
\value{
TRUE/FALSE
}
\description{
Simple function that returns TRUE/FALSE if "federal" is detected, case insensitive.
}
\examples{
is_federal("FedeRal")
is_federal("test")

}
