% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_hifld_id.R
\name{is_hifld_id}
\alias{is_hifld_id}
\title{Returns TRUE if the given ID is a valid HIFLD ID}
\usage{
is_hifld_id(id, hifld_data = NULL, na_is_na = F)
}
\arguments{
\item{id}{integer or character string of the ID to check}

\item{hifld_data}{data frame with HIFLD information}

\item{na_is_na}{logical, whether NA should return NA (vs. FALSE)}
}
\value{
logical, TRUE if the given ID is a valid HIFLD ID
}
\description{
Returns TRUE if the given ID is a valid HIFLD, includes a parameter to 
return FALSE vs. NA if the given ID is NA.
}
\examples{
is_hifld_id(10002598) 
is_hifld_id(c(10002598, 123, NA))
}
