% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_not_na.R
\name{last_not_na}
\alias{last_not_na}
\title{Replaces NA in a vector with the previous most non NA value}
\usage{
last_not_na(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with NAs replaced with preceding value
}
\description{
Replaces NA in a vector with the previous most non NA value. If no preceding
non NA values then keep as NA.
}
\examples{
last_not_na(c(NA,1, NA, 1:3, NA))

}
