% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_remote_data.R
\name{list_remote_data}
\alias{list_remote_data}
\title{List remote data locations for Behind Bars Data}
\usage{
list_remote_data(folder, scraper_name = NULL, dates_only = FALSE)
}
\arguments{
\item{folder}{character, a single character indicating the folder to pull
file locations from. Must be one of "raw_files", "log_files", or
"extracted_data"}

\item{scraper_name}{character, a character string indicating a particular
scraper to pull data for. Defaults to NULL which pulls all file locations}

\item{dates_only}{logical, only return the dates rather than full file name.
Should only be used when scraper_name is not null}
}
\value{
character vector of file location urls
}
\description{
List remote data locations for Behind Bars Data for either raw_files,
log_files, or extracted_data. a scraper_name may also be provided to pull
only a specific subset of locations.
}
\examples{
list_remote_data("log_files")
list_remote_data("raw_files", "lasd")
list_remote_data("raw_files", "lasd", TRUE)

}
