% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_na_rm.R
\name{max_na_rm}
\alias{max_na_rm}
\title{A re-coding of the max function which returns NA if all values are NA}
\usage{
max_na_rm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of length 1
}
\description{
A re-coding of the max function which returns NA if all values are NA.
Normally max(na.rm=TRUE) will return -Inf if all values are NA.
}
\examples{
max_na_rm(c(1:2, NA))
max_na_rm(c(NA, NA))

}
