% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_facility_info.R
\name{merge_facility_info}
\alias{merge_facility_info}
\title{A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations}
\usage{
merge_facility_info(dat)
}
\arguments{
\item{dat}{Scraped/historical data with columns Name, State, jurisdiction}
}
\value{
data set with cleaned facility name column, "Name", from crosswalk on GitHub
}
\description{
A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations.
Cleans federal and non-federal facilities in separate processes, which requires a column named "jurisdiction"
to indicate federal/nonfederal
}
