% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTable.R
\name{parse_response}
\alias{parse_response}
\title{Parse API response from extractable}
\usage{
parse_response(server_resp)
}
\arguments{
\item{server_resp}{response from httr}
}
\value{
extracted list results from json
}
\description{
Parse API response from extractable
}
