% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fac_trend.R
\name{plot_fac_trend}
\alias{plot_fac_trend}
\title{Plot a single facility metric}
\usage{
plot_fac_trend(
  fac_name,
  state,
  metric,
  scrape_df = NULL,
  plot_days = NULL,
  annotate = F
)
}
\arguments{
\item{fac_name}{character string of the facility to plot}

\item{state}{character string of the facility's state}

\item{metric}{character string of the metric to plot}

\item{scrape_df}{data frame with scraped data}

\item{plot_days}{integer, number of days to plot}

\item{annotate}{logical, whether to include text annotations for most recent value}
}
\value{
ggplot object
}
\description{
Creates time series plots showing the trend for a single facility for a given
metric, formatted according to the style guide.
}
\examples{
plot_fac_trend("Los Angeles Jails", "California", "Residents.Active", scrape_df)

}
