% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recent_fac_increases.R
\name{plot_recent_fac_increases}
\alias{plot_recent_fac_increases}
\title{Plot facilities with recent COVID-19 metric increases}
\usage{
plot_recent_fac_increases(
  scrape_df = NULL,
  metric = "Residents.Confirmed",
  plot_days = NULL,
  delta_days = 7,
  num_fac = 4,
  annotate = F
)
}
\arguments{
\item{scrape_df}{data frame with scraped data}

\item{metric}{character string of the metric to plot}

\item{plot_days}{integer, number of days to plot}

\item{delta_days}{integer, number of days to calculate increase over}

\item{num_fac}{integer, number of facilities to plot}

\item{annotate}{logical, whether to include text annotations for most recent value}
}
\value{
ggplot object
}
\description{
Creates time series plots showing trends for facilities with the steepest recent 
increases in a given metric, formatted according to the style guide.
}
\examples{
plot_recent_fac_increases()
plot_recent_fac_increases(metric = "Residents.Deaths", plot_days = 30, num_fac = 3)

}
