% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTable.R
\name{process_file}
\alias{process_file}
\title{Post image for extraction from API endpoint}
\usage{
process_file(api_key, filepath)
}
\arguments{
\item{api_key}{character of extractable api key}

\item{filepath}{character to image filepath to post}
}
\description{
Post image for extraction from API endpoint
}
