% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aggregate_pop_data.R
\name{read_aggregate_pop_data}
\alias{read_aggregate_pop_data}
\title{Read in statewide population data}
\usage{
read_aggregate_pop_data()
}
\value{
data frame with population data
}
\description{
Reads in statewide-aggregated anchored population data with rows for Federal,
ICE and 51 state agencies.
}
\examples{
read_aggregate_pop_data()

}
