% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cpp_data.R
\name{read_cpp_data}
\alias{read_cpp_data}
\title{Read the latest COVID Prison Project State Data}
\usage{
read_cpp_data()
}
\value{
data frame with COVID Prison Project results
}
\description{
Reads the COVID Prison Project dataset and converts variable names to be 
comparable to the UCLA dataset
}
\examples{
\dontrun{
read_cpp_data()
}
}
