% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fac_crosswalk.R
\name{read_fac_crosswalk}
\alias{read_fac_crosswalk}
\title{Read in UCLA Prison and Jail Facility Data Crosswalk}
\usage{
read_fac_crosswalk()
}
\value{
data frame with facility crosswalk data
}
\description{
Reads in facility crosswalk data. NB:
}
