% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fac_info.R
\name{read_fac_info}
\alias{read_fac_info}
\title{Read in UCLA Prison and Jail Facility Data}
\usage{
read_fac_info()
}
\value{
data frame with facility metadata
}
\description{
Reads in facility metadata
}
