% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fac_spellings.R
\name{read_fac_spellings}
\alias{read_fac_spellings}
\title{Read in UCLA Prison and Jail Facility Data}
\usage{
read_fac_spellings()
}
\value{
data frame with facility identifiers and variations on those names
}
\description{
Reads in facility alternative name spellings
}
