% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hifld_data.R
\name{read_hifld_data}
\alias{read_hifld_data}
\title{Read in HIFLD Facility level Data}
\usage{
read_hifld_data()
}
\value{
data frame with facility info
}
\description{
Reads in data sheet from HIFLD on facility information including population
}
