% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpap_data.R
\name{read_mpap_data}
\alias{read_mpap_data}
\title{Read the latest Marshall Project State Data}
\usage{
read_mpap_data(all_dates = FALSE, window = 14)
}
\arguments{
\item{all_dates}{return all historical data from MP/AP}

\item{window}{integer, the day range of acceptable data to pull from if
all_dates is false}
}
\value{
data frame with MP/AP results
}
\description{
Reads the MP/AP dataset and converts variable names to be comparable to the
UCLA dataset
}
\examples{
\dontrun{
read_mpap_data()
}
}
