% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mpap_pop_data.R
\name{read_mpap_pop_data}
\alias{read_mpap_pop_data}
\title{Get Feb 20, 2020 population data for state prisons from the MP}
\usage{
read_mpap_pop_data()
}
\value{
data frame with jurisdiction level pop data from MP
}
\description{
Reads the MP dataset and for each prison jurisdiction pulls the staff and
resident population data closest to Feb 20, 2020.
}
\examples{
\dontrun{
read_staff_popfeb20()
}
}
