% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop_data.R
\name{read_pop_data}
\alias{read_pop_data}
\title{Read in UCLA Population Data with cleaned names}
\usage{
read_pop_data()
}
\value{
data frame with population data
}
\description{
Reads in population data
}
