% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_staff_popfeb20.R
\name{read_staff_popfeb20}
\alias{read_staff_popfeb20}
\title{Get Feb 20, 2020 population data for applicable rows in the fac_data}
\usage{
read_staff_popfeb20()
}
\value{
data frame with population data for staff by facility
}
\description{
Reads the UCLA dataset and for each facility pulls the staff population data
closest to Feb 20, 2020 in the UCLA historical data.
}
\examples{
\dontrun{
read_staff_popfeb20()
}
}
