% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vera_pop.R
\name{read_vera_pop}
\alias{read_vera_pop}
\title{Read in Vera Prison population estimates}
\usage{
read_vera_pop(updated = TRUE)
}
\arguments{
\item{updated}{logical, use updated population estimates from March where
available}
}
\value{
data frame with population estimates
}
\description{
Reads in an estimate of Vera prison population estimates
}
