% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_cols.R
\name{reorder_cols}
\alias{reorder_cols}
\title{Re-orders columns to respect structure of daily scraped data.}
\usage{
reorder_cols(data, add_missing_cols = TRUE, rm_extra_cols = FALSE)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{add_missing_cols}{logical, add missing columns as all NA rows}

\item{rm_extra_cols}{logical, remove extra columns}
}
\value{
re-ordered data frame
}
\description{
Reorders columns in a data frame to agree with the existing structure.
Throws a warning for missing columns and optionally adds them in as all NA rows.
Throws a warning for extra columns and optionally keeps them at the end of the dataframe.
}
\examples{
test_data <- tibble::tibble("jurisdiction" = "jail",
                    "City" = "Los Angeles")
reorder_cols(test_data)

}
