% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTable.R
\name{retrieve_result}
\alias{retrieve_result}
\title{Retrieve result from API endpoint}
\usage{
retrieve_result(api_key, job_id)
}
\arguments{
\item{api_key}{character of extractable api key}

\item{job_id}{character of job id to pull}
}
\value{
raw extracted results
}
\description{
Retrieve result from API endpoint
}
