% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_sum.R
\name{rolling_sum}
\alias{rolling_sum}
\title{A function to add in a rolling sum estiomate column by groups}
\usage{
rolling_sum(
  .data,
  ...,
  .col = "Residents.Confirmed",
  .window = 14,
  .new_col = "Residents.Active.Estimate"
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr). Must have a date column of class
date}

\item{...}{one or more unquoted expressions separated by commas. Variable
names can be used as if they were positions in the data frame,
so expressions like x:y can be used to select a range of variable. These
columns should uniquely identify rows in the data}

\item{.col}{character, the column to perform the rolling sum on}

\item{.window}{integer, the window of retrospective days for rolling sum}

\item{.new_col}{character, the new column name to be added on}
}
\value{
data frame with new column with rolling sum added
}
\description{
Adds in a rolling sum calculation to a data frame which can be used as a way
to estimate active cases from confirmed cases.
}
\examples{
read_scrape_data(all_dates = TRUE, state = "North Carolina") \%>\%
    filter(Name == "NORTH CAROLINA CORRECTIONAL INSTITUTION FOR WOMEN") \%>\%
    rolling_sum(Name) \%>\%
    ggplot(aes(
        x = Date, y = Residents.Active.Estimate, color = Name, fill = Name)) +
    geom_line(size = 1.5) +
    geom_area(alpha = .5) +
    theme_behindbars() +
    scale_color_bbdiscrete() +
    scale_fill_bbdiscrete() +
    theme(legend.position = "none") +
    labs(y = "Estimated Active Cases") +
    ggtitle("Monitoring Facility Outbreaks")

}
