% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_na_rm.R
\name{sum_na_rm}
\alias{sum_na_rm}
\title{A re-coding of the sum function which returns NA if all values are NA}
\usage{
sum_na_rm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of length 1
}
\description{
A re-coding of the sum function which returns NA if all values are NA.
Normally sum(na.rm=TRUE) will return 0 if all values are NA.
}
\examples{
sum_na_rm(c(1:2, NA))
sum_na_rm(c(NA, NA))

}
