% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_behindbars.R
\name{theme_behindbars}
\alias{theme_behindbars}
\alias{scale_color_bbdiscrete}
\alias{scale_fill_bbdiscrete}
\alias{scale_color_bbcontinous}
\alias{scale_fill_bbcontinous}
\title{Theme and colors for behind bars plots}
\usage{
theme_behindbars()

scale_color_bbdiscrete()

scale_fill_bbdiscrete()

scale_color_bbcontinous()

scale_fill_bbcontinous()
}
\description{
a set of themes and colors for behind bars plots
}
\examples{
library(ggplot2)
iris \%>\%
    ggplot(aes(x = Petal.Width, y = Petal.Length, color = Species)) +
    geom_point() +
    theme_behindbars() +
    scale_color_bbdiscrete()

iris \%>\%
    ggplot(aes(x = Petal.Width, fill = Species)) +
    geom_density() +
    theme_behindbars() +
    scale_fill_bbdiscrete()

iris \%>\%
    ggplot(aes(x = Petal.Width, y = Petal.Length, color = Petal.Width)) +
    geom_point() +
    theme_behindbars() +
    scale_color_bbcontinous()
}
