% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_state.R
\name{translate_state}
\alias{translate_state}
\title{Converts state abbreviations to state names}
\usage{
translate_state(x, reverse = FALSE)
}
\arguments{
\item{x}{character vector of state abbreviations or names}

\item{reverse}{logical, reverse translation from full name to abbreviation}
}
\value{
character vector of state names
}
\description{
Converts state abbreviations to state names or does the reverse
}
\examples{
translate_state(c("CA", "NM", "DC"))

}
