% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_sum_na_rm.R
\name{vector_sum_na_rm}
\alias{vector_sum_na_rm}
\title{A re-coding of the sum function which returns NA if all values are NA}
\usage{
vector_sum_na_rm(...)
}
\arguments{
\item{...}{numeric vectors of equal length}
}
\value{
numeric vector of length 1
}
\description{
A re-coding of the sum function which returns NA if all values are NA.
This function sums vectors together.
}
\examples{
sum_na_rm(c(1:2, NA))
sum_na_rm(c(NA, NA))

}
