% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wayback_archives.R
\name{wayback_archives}
\alias{wayback_archives}
\title{Given a URL returns the url, timestamp, and statuscodes for wayback machine}
\usage{
wayback_archives(url)
}
\arguments{
\item{url}{character, url to check wayback archives for}
}
\value{
data frame like object with columns for original url, timestamp,
statuscode and url
}
\description{
given a URL returns the time stamped url, timestamp, and statuscodes
for wayback machine archived versions of the website.
}
\examples{
\dontrun{
wayback_archives("https://doccs.ny.gov/doccs-covid-19-report")
}

}
