% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aggregate_counts.R
\name{calc_aggregate_counts}
\alias{calc_aggregate_counts}
\title{Aggregate UCLA and MP data to get a more recent accurate count of COVID variables}
\usage{
calc_aggregate_counts(
  window = 31,
  ucla_only = FALSE,
  state = FALSE,
  collapse_vaccine = TRUE,
  all_dates = FALSE,
  week_grouping = TRUE
)
}
\arguments{
\item{window}{integer, the day range of acceptable data to pull from, ignored
if all dates is true}

\item{ucla_only}{logical, only consider data from UCLA}

\item{state}{logical, return state level data}

\item{collapse_vaccine}{logical, combine vaccine variables for more
intuitive comparisons}

\item{all_dates}{logical, get time series data rather than just latest counts}

\item{week_grouping}{logical, use weekly grouping for past data? else monthly}
}
\value{
data frame with aggregated counts at state or national level
}
\description{
Reads the UCLA and MP/AP dataset aggregates counts for states for most recent
data within a given window and reports either state level data or national
data. States include values for the 50 state DOCs, Federal for BOP prisons,
ICE detention centers, and incarcerated individuals under the administration
of the District of Columbia DOC. If both UCLA and MP report a
value for a state the larger value for is taken.
}
\examples{
\dontrun{
calc_aggregate_counts()
}
calc_aggregate_counts(state = TRUE, all_dates = TRUE)
}
