% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aggregate_counts.R
\name{calc_aggregate_counts}
\alias{calc_aggregate_counts}
\title{Aggregate UCLA and MP data to get a more recent accurate count of COVID variables}
\usage{
calc_aggregate_counts(window = 31, ucla_only = FALSE, state = FALSE)
}
\arguments{
\item{window}{integer, the day range of acceptable data to pull from}

\item{ucla_only}{logical, only consider data from UCLA}

\item{state}{logical, return state level data}
}
\value{
data frame with aggregated counts at state or national level
}
\description{
Reads the UCLA and MP/AP dataset aggregates counts for states for most recent
data within a given window and reports either state level data or national
data. States include values for the 50 states, Federal for BOP prisons, and
District of Columbia for prison in the capitol.
}
\examples{
\dontrun{
calc_aggregate_counts()
}
}
