% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_facility_name.R
\name{clean_facility_name}
\alias{clean_facility_name}
\title{A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations}
\usage{
clean_facility_name(dat)
}
\arguments{
\item{dat}{Scraped/historical data with columns Name and State, at the very least}
}
\value{
data set with cleaned facility name column, "Name", from crosswalk on GitHub
}
\description{
A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations.
Cleans federal and non-federal facilities in separate processes, in order to use "State" to merge or not
}
