% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_facility_name.R
\name{clean_facility_name}
\alias{clean_facility_name}
\title{A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations}
\usage{
clean_facility_name(dat, alt_name_xwalk = FALSE, debug = FALSE)
}
\arguments{
\item{dat}{Scraped/historical data with columns Name and State, at the very least}

\item{alt_name_xwalk}{Optional parameter provides an alternative facility name crosswalk}

\item{debug}{Boolean whether to include additional columns geneated during the merging process}
}
\value{
data set with cleaned columns, "Name", "Facility.ID", and "Jurisdiction" from fac_spellings
}
\description{
A facility name cleaning function. Uses GitHub facility name cross-walk to find all possible name variations.
Cleans federal and non-federal facilities in separate processes, in order to use "State" to merge or not.
If no match is found in the crosswalk, both Facility.ID and Jurisdiction will equal NA in the resulting output.
}
\details{
This function assumes that if a faceility is federal, then vars "Facility" OR "State" OR "Jurisdiction" will
contain the word federal (case-insensitive).

For non-federal entities, the data returned gets "Jurisdiction" from fac_spellings

For federal + immigration entities, the data returned gets "Jurisdiction" and "State" from fac_spellings

When there is not enough info to get a match, both "Jurisdiction" and "Facility.ID" will be returned as NA
}
\examples{
clean_facility_name(
    tibble(Name = "BULLOCK CORRECTIONAL FACILITY", State = "Alabama", Jurisdiction = "state"))

clean_facility_name(
    tibble(Name = "LEE USP", State = "Federal", Facility = "prison"))

}
