% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_facility_info.R
\name{merge_facility_info}
\alias{merge_facility_info}
\title{Merges informative facility information such as capacity, geographic information, and population.}
\usage{
merge_facility_info(dat)
}
\arguments{
\item{dat}{Scraped/historical data with column "Facility.ID"}
}
\value{
data set with facility information columns
}
\description{
Merges our facility dataset with detailed descriptive information on each entity in our
dataset (e.g. type of entity, capacity, population, geographic information, etc.).
Each row represents a unique entity.
Data source here: https://github.com/uclalawcovid19behindbars/facility_data/data/fac_data.csv
}
\examples{
merge_facility_info(
    tibble(Name = "BULLOCK CORRECTIONAL FACILITY", State = "Alabama", Jurisdiction = "state", Facility.ID = 7))


}
