% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fac_trend.R
\name{plot_fac_trend}
\alias{plot_fac_trend}
\title{Plot a single facility metric}
\usage{
plot_fac_trend(
  fac_name,
  state,
  metric,
  scrape_df = NULL,
  plot_days = NULL,
  area_plot = F,
  annotate = F,
  auto_label = F
)
}
\arguments{
\item{fac_name}{character string of the facility to plot}

\item{state}{character string of the facility's state}

\item{metric}{character string of the metric to plot}

\item{scrape_df}{data frame with scraped data}

\item{plot_days}{integer, number of days to plot}

\item{area_plot}{logical, whether to create an area plot (vs. a line plot)}

\item{annotate}{logical, whether to include text annotations for most recent value}

\item{auto_label}{logical, whether to label plot title, subtitle, axes based on behindbarstools::get_metric_description}
}
\value{
ggplot object
}
\description{
Creates time series plots showing the trend for a single facility for a given
metric, formatted according to the style guide. Returns a ggplot object, so
additional layers can be added, per the examples.
}
\examples{
\dontrun{
plot_fac_trend("Los Angeles Jails", "California", "Residents.Active")
}

\dontrun{
plot_fac_trend("Los Angeles Jails", "California", "Residents.Active") +
    scale_x_date(date_labels = "\%m/\%y") +
    labs(x = "Cases") +
    ylim(0, 150)
}

}
