% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hifld_data.R
\name{read_hifld_data}
\alias{read_hifld_data}
\title{Read in HIFLD Facility level Data}
\usage{
read_hifld_data()
}
\value{
data frame with facility info and coordinates
}
\description{
Reads in data from HIFLD on facility information including population, adds
coordinates (latitude and longitude) for the centroid of each facility, replaces
-999 population and capacity values with NA.
}
\examples{
\dontrun{
read_hifld_data()
}
}
