% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scrape_data.R
\name{read_scrape_data}
\alias{read_scrape_data}
\title{Read data extracted by webscraper}
\usage{
read_scrape_data(
  all_dates = FALSE,
  window = 31,
  coalesce = TRUE,
  coalesce_pop = TRUE,
  drop_na_covid = TRUE,
  debug = FALSE,
  state = NULL
)
}
\arguments{
\item{all_dates}{logical, get all data from all dates recorded by webscraper}

\item{window}{int, if all dates is false how far back to look for values from
a given facility to populate values until value should be NA}

\item{coalesce}{logical, collapse common facilities into single row}

\item{coalesce_pop}{logical, collapse population data based on the date window}

\item{drop_na_covid}{logical, drop rows missing all COVID variables}

\item{debug}{logical, print debug statements on number of rows maintained in}

\item{state}{character vector, states to limit data to
cleaning process}
}
\value{
character vector of state names
}
\description{
Reads either time series or latest data from the web scraper runs
}
\examples{
\dontrun{
read_scrape_data(all_dates = FALSE)
}
read_scrape_data(all_dates = TRUE, state = "Wyoming")

}
