% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_sum_na_rm.R
\name{vector_sum_na_rm}
\alias{vector_sum_na_rm}
\title{A re-coding of the sum function which returns NA if all values are NA}
\usage{
vector_sum_na_rm(...)
}
\arguments{
\item{...}{numeric vectors of equal length}
}
\value{
numeric vector of length equal to provided vectors
}
\description{
A re-coding of the sum function which returns NA if all values are NA.
This function sums vectors together.
}
\examples{
tibble(x = c(NA, NA, NA), y = c(NA, 2, NA), z = c(1, NA, NA)) \%>\%
   mutate(new = vector_sum_na_rm(x, y, z))
}
